Move connector to its own module
